+++ okitweb/static/okit/views/compartment/js/resources/oke_cluster.js
@@ -7,13 +7,11 @@ console.debug('Loaded Compartment OKE View Javascript');
 /*
 ** Define Compartment View Artifact Class
  */
-class OkeClusterView extends OkitCompartmentArtefactView {
+class OkeClusterView extends OkitContainerCompartmentArtefactView {
     constructor(artefact = null, json_view) {
         super(artefact, json_view);
     }
 
-    // get parent_id() {return this.artefact.vcn_id;}
-    // get parent() {return this.getJsonView().getVirtualCloudNetwork(this.parent_id);}
     get parent_id() {
         let vcn = this.getJsonView().getVirtualCloudNetwork(this.artefact.vcn_id);
         if (vcn && vcn.compartment_id === this.artefact.compartment_id) {
@@ -22,17 +20,22 @@ class OkeClusterView extends OkitCompartmentArtefactView {
             return this.artefact.compartment_id;
         }
     }
-    get parent() {return this.getJsonView().getVirtualCloudNetwork(this.parent_id) ? this.getJsonView().getVirtualCloudNetwork(this.parent_id) : this.getJsonView().getCompartment(this.parent_id);}
     // ---- Okit View Functions
 
     /*
     ** SVG Processing
      */
+    /*
+    ** Property Sheet Load function
+    */
+    newPropertiesSheet() {
+        this.properties_sheet = new OkeClusterProperties(this.artefact)
+    }
 
     /*
     ** Property Sheet Load function
      */
-    loadProperties() {
+    loadProperties1() {
         let me = this;
         $(jqId(PROPERTIES_PANEL)).load("propertysheets/oke_cluster.html", () => {
             let service_lb_subnet_select = d3.select(d3Id('service_lb_subnet_ids'));
@@ -327,6 +330,13 @@ class OkeClusterView extends OkitCompartmentArtefactView {
     }
 
     /*
+    ** Child Type Functions
+     */
+    getTopArtifacts() {
+        return [NodePool.getArtifactReference()];
+    }
+
+    /*
     ** Static Functionality
      */
     static getArtifactReference() {
@@ -334,7 +344,7 @@ class OkeClusterView extends OkitCompartmentArtefactView {
     }
 
     static getDropTargets() {
-        return [VirtualCloudNetwork.getArtifactReference()];
+        return [Compartment.getArtifactReference(), VirtualCloudNetwork.getArtifactReference()];
     }
 
 }